Changing underlying storage to any `PinnedVec`
* Instead of hard-coded `SplitVec`, the concurrent bag now wraps any `PinnedVec`.
* `capacity` method is added.
* `Debug` is implemented.
* Benchmarks added and reported.
* `AtomicUsize` is used for capacity rather than `usize`. This update does not lead to diminished performance.
* Memory orderings of atomic operations are strengthened.
* `UnsafeCell` is used for the pinned vector, rather than direct cast.
* Tests extended.

# Inside the push

In the prior version, growth of the bag required two operations:
* setting length of the pinned vector to its capacity, and
* growing the pinned vector to its next size.

The reason the first operation is required is due to the fact that the second operation can only succeed when vector is at its capacity; i.e., `len == capacity`.

Having to make two mutations is not good for heavy load situations. `grow_to` method, on the other hand, does not require the pinned vector to be at its length, and hence, we do not need to make the `set_len` call. Furthemore, it will be required for a possible `extend` implementation.

# Also

* Errors are organized.
* Memory orderings are strengthened.