concurrency-model-revised
# Concurrency Model Revision

With this major refactoring, safety issues are fixed to achieve a miri safe version for all the tests defined in this crate.

The refactoring consists of two main parts: downstream changes in pinned vector's concurrency support and replacing Option<T> with T and AtomicBool vector.

## PinnedVec Support for Concurrency

In version 2, PinnedVec grew with new methods to support concurrent data structures. However, this caused problems since these exposed methods were often unsafe, and further, they were not directly useful for the pinned vector consumers except for concurrent data structures wrapping a pinned vector. Furthermore, they are alien to a regular vector interface that we are used to using.

In version 3, a second trait called `ConcurrentPinnedVec` is defined. All useful methods related with concurrent programming are moved to this trait. This trait has an associated type defining the underlying pinned vector type. It can be turned into the pinned vector.

Finally, `IntoConcurrentPinnedVec` trait is defined. A pinned vector implementing this trait can be turned into a `ConcurrentPinnedVec`. As explained above, it can be converted back to the pinned vector.

In this release `ConcurrentVe` wraps a `ConcurrentPinnedVec` rather than a `PinnedVec`.

## Concurrency State

The main goal of ConcurrentVec is to enable safe reading while allowing it to efficiently grow. In version 1, it was attempted to achieve this through wrapping the elements by an Option. Although this turned out to pass exhaustive tests, it had two problems. Firstly, it is not best to have a vector of Option<T> rather than that of T's, since the concurrent collection allows to transform `into_inner`. In version 2, underlying concurrent pinned vector stores elements as T, and hence, we can directly get the vector of T's. Secondly, version 1 failed miri tests for the tests where we read & grow concurrently. Having a parallel concurrent vector of atomic bools, the race condition is eliminated in version 2.